requirements.txt
# BioGestão 360 — Dependências do projeto (versão 4.2)
# Instalação: pip install -r requirements.txt

streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.24.0
matplotlib>=3.7.0
qrcode[pil]>=7.0.0
pillow>=10.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
numpy>=1.24.0
pyotp>=2.9.0
